release: version 0.0.6

+++ lib/features/lost_objects/domain/usecases/delete_lost_object_usecase.dart
@@ -9,7 +9,11 @@ class DeleteLostObjectUseCase {
   Future<void> call({
     required String requesterId,
     required LostObject object,
-  }) {
+  }) async {
+    if (!object.canBeDeleted) {
+      throw StateError(object.deletionBlockedReason);
+    }
+
     return _repository.deleteLostObject(
       requesterId: requesterId,
       objectId: object.id,
